requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
pytgcalls==3.0.0.dev31
motor==3.3.2
yt-dlp>=2024.1.1
google-generativeai>=0.5.0
python-dotenv>=1.0.0
spotipy>=2.23.0
aiohttp>=3.9.0
